+++ backdev/src/index.ts
@@ -22,10 +22,10 @@ import {
     // No auto order
     autoOrder: "None",
     renderOptions: {
-      bitmapViewThreshold: Zoom.Divide4,
-      gridDisplayThreshold: Zoom.Divide2,
-      textureDisplayThreshold: Zoom.Divide2,
-      textDisplayThreshold: Zoom.Base,
+      bitmapViewThreshold: Zoom.Divide16,
+      gridDisplayThreshold: Zoom.Base,
+      textureDisplayThreshold: Zoom.Base,
+      textDisplayThreshold: Zoom.Multiply2,
     },
   });
 
@@ -111,12 +111,12 @@ import {
 
   for (let i = 0; i < 10000; i++) {
     // Randomize the array for each iteration
-    for (let iY = 0; iY < size.y; iY++) {
-      for (let iX = 0; iX < size.x; iX++) {
-        myArray[iY][iX] = Math.random();
-        myArrayTop[iY][iX] = Math.random();
-      }
-    }
+    // for (let iY = 0; iY < size.y; iY++) {
+    //   for (let iX = 0; iX < size.x; iX++) {
+    //     myArray[iY][iX] = Math.random();
+    //     myArrayTop[iY][iX] = Math.random();
+    //   }
+    // }
 
     // Another way to get keyboard events.
     // KeyCode is bind on the physical position of the key.
@@ -131,7 +131,7 @@ import {
       // Options
       {
         // Wait for 1000ms before continuing.
-        waitFor: 0,
+        waitFor: 50,
         // Wait for the array to be rendered before continuing.
         isSync: true,
         mapper: customMapper,
@@ -140,21 +140,21 @@ import {
       }
     );
 
-    await mem.log2d(
-      // Array unique id
-      "my_array_id_2",
-      // Array reference
-      myArrayTop,
-      // Options
-      {
-        // Wait for 1000ms before continuing.
-        waitFor: 33,
-        // Wait for the array to be rendered before continuing.
-        isSync: true,
-        mapper: customMapperTop,
-        output: customOutput,
-        position: { x: 0, y: 0 },
-      }
-    );
+    // await mem.log2d(
+    //   // Array unique id
+    //   "my_array_id_2",
+    //   // Array reference
+    //   myArrayTop,
+    //   // Options
+    //   {
+    //     // Wait for 1000ms before continuing.
+    //     waitFor: 33,
+    //     // Wait for the array to be rendered before continuing.
+    //     isSync: true,
+    //     mapper: customMapperTop,
+    //     output: customOutput,
+    //     position: { x: 0, y: 0 },
+    //   }
+    // );
   }
 })();