+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2024-11-04
+
+### Changed
+
+- Texture Rendering: Floor texture position in order to have clean render.
+- Texture Rendering: Fix blur on window resize.
+- Separate rendering by layers.
+
 ## [0.3.0] - 2024-11-01
 
 ### Added